For EXP-2012: Move tests for updating into their own file
