+++ components/nimbus/src/tests/mod.rs
@@ -11,6 +11,7 @@ mod test_lib_bw_compat;
 mod test_lib_schema_deserialization;
 mod test_persistence;
 mod test_sampling;
+mod test_updating;
 mod test_versioning;
 
 mod client {
